4.Faturamento/requirements.txt
numpy==1.26.4
streamlit==1.37.1
pandas==2.2.2
plotly-express==0.4.1
openpyxl==3.1.2
matplotlib==3.8
scikit-learn==1.5.1
fdb == 2.0.2

